feat: Add download endpoint for inputs
This commit adds a new endpoint `/v1/inputs/download` to the input route, which allows users to download inputs in Excel format. The `download` method is implemented in the `InputController` and `InputService` classes.

The Excel file contains the following columns:
- Name
- Input Type
- Code
- Unit of Measure
- Cost
- Supplier
- VAT
- Price

The commit also includes necessary changes to the OpenAPI documentation for the new endpoint.